# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/google": {
      "post": {
        "operationId": "AuthController_google",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthGoogleRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthGoogleResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Google id_token invalido o no verificable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "summary": "Login con Google id_token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales invalidas o usuario inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "summary": "Login normal con email y password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLogoutResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rimac_refresh": []
          }
        ],
        "summary": "Revocar la sesion indicada por la cookie HttpOnly",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLogoutResponseDocDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revocar todas las sesiones activas del usuario",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRefreshResponseDocDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalido, expirado o usuario inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rimac_refresh": []
          }
        ],
        "summary": "Renovar access token con cookie HttpOnly",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRegisterResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "409": {
            "description": "El email ya esta registrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "summary": "Registro con email y password",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bus Impl API",
    "description": "API de negocio actualizada a 2026",
    "version": "2026.06",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorResponseDocDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "AuthRegisterRequestDocDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 160,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "full_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "full_name",
          "password"
        ]
      },
...
      "ApiValidationErrorResponseDocDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Campo del request que fallo la validacion."
                },
                "message": {
                  "type": "string",
                  "description": "Mensaje legible del error de validacion."
                }
              },
              "required": [
                "field",
                "message"
              ]
            }
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "AuthLoginRequestDocDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 160,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthLoginResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "token_type": {
            "type": "string",
            "const": "Bearer",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "access_token": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "expires_in": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "email": {
                "type": "string",
                "maxLength": 160,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "full_name": {
                "type": "string",
                "minLength": 2,
                "maxLength": 120
              },
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "operador",
                  "auditor",
                  "soporte"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive",
                  "blocked"
                ]
              },
              "provider": {
                "type": "string",
                "enum": [
                  "local",
                  "google"
                ]
              },
              "google_sub": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "required": [
              "id",
              "email",
              "full_name",
              "role",
              "status",
              "provider",
              "google_sub",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "token_type",
          "access_token",
          "expires_in",
          "user"
        ]
      },
      "AuthGoogleRequestDocDto": {
        "type": "object",
        "properties": {
          "id_token": {
            "type": "string",
            "minLength": 20
          }
        },
        "required": [
          "id_token"
        ]
      },
      "AuthGoogleResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "token_type": {
            "type": "string",
            "const": "Bearer",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "access_token": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "expires_in": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "email": {
                "type": "string",
                "maxLength": 160,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "full_name": {
                "type": "string",
                "minLength": 2,
                "maxLength": 120
              },
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "operador",
                  "auditor",
                  "soporte"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive",
                  "blocked"
                ]
              },
              "provider": {
                "type": "string",
                "enum": [
                  "local",
                  "google"
                ]
              },
              "google_sub": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "required": [
              "id",
              "email",
              "full_name",
              "role",
              "status",
              "provider",
              "google_sub",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "token_type",
          "access_token",
          "expires_in",
          "user"
        ]
      },
      "AuthRefreshRequestDocDto": {
        "type": "object",
        "properties": {
          "root": {
            "default": {},
            "type": "object",
            "properties": {},
            "additionalProperties": false,
            "x-nestjs_zod-unwrap-root": true,
            "x-nestjs_zod-self-required": false
          }
        }
      },
      "AuthRefreshResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "token_type": {
            "type": "string",
            "const": "Bearer",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "access_token": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "expires_in": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "email": {
                "type": "string",
                "maxLength": 160,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "full_name": {
                "type": "string",
                "minLength": 2,
                "maxLength": 120
              },
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "operador",
                  "auditor",
                  "soporte"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive",
                  "blocked"
                ]
              },
              "provider": {
                "type": "string",
                "enum": [
                  "local",
                  "google"
                ]
              },
              "google_sub": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "required": [
              "id",
              "email",
              "full_name",
              "role",
              "status",
              "provider",
              "google_sub",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "token_type",
          "access_token",
          "expires_in",
          "user"
        ]
      },
      "AuthLogoutResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "revoked_sessions": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "revoked_sessions"
        ]
      },
...
    }
  }
}